Fix to table downloads / added cacheBuilt flag
Seemed to have broken table downloads in a recent commit. Should now be
fixed (and possibly faster).

Running sourcery_build_db now adds a cacheBuilt flag to the sourceCollection,
which is initally set to 0. After each object is added in
sourcery_build_cache, this flag is set to 1. Allows the latter script to
restart quickly if anything goes wrong (e.g., download fails).